refactor(cdconc): format & cleanup
chore(cdconc): remove comments regarding TypeCheck3
