+++ cddiff/src/main/java/de/monticore/cdconcretization/association/IAssocSideCompleter.java
@@ -4,7 +4,8 @@ import de.monticore.cdassociation._ast.ASTCDAssocSide;
 import de.monticore.cdconcretization.CompletionException;
 
 /**
- * Completes a concrete association side such that it conforms to a given reference association side.
+ * Completes a concrete association side such that it conforms to a given reference association
+ * side.
  */
 public interface IAssocSideCompleter {
