+++ cddiff/src/main/java/de/monticore/cdconcretization/cd/CDCompletionContext.java
@@ -6,7 +6,6 @@ import de.monticore.cdbasis._ast.ASTCDCompilationUnit;
 import de.monticore.cdbasis._ast.ASTCDType;
 import de.monticore.cdconformance.CDConfParameter;
 import de.monticore.cdmatcher.MatchingStrategy;
-
 import java.util.Set;
 
 /**
