+++ cddiff/src/main/java/de/monticore/cdconcretization/cd/type/BaseTypeInCDCompleter.java
@@ -2,6 +2,7 @@ package de.monticore.cdconcretization.cd.type;
 
 import de.monticore.cdbasis._ast.ASTCDClass;
 import de.monticore.cdbasis._ast.ASTCDDefinition;
+import de.monticore.cdconcretization.CompletionException;
 import de.monticore.cdconcretization.cd.CDCompletionContext;
 import de.monticore.cdinterfaceandenum._ast.ASTCDEnum;
 import de.monticore.cdinterfaceandenum._ast.ASTCDInterface;
@@ -11,7 +12,8 @@ public class BaseTypeInCDCompleter extends AbstractTypeInCDCompleter {
 
   @Override
   protected void completeType(
-      ASTCDDefinition concreteCD, ASTCDClass referenceType, CDCompletionContext context) {
+      ASTCDDefinition concreteCD, ASTCDClass referenceType, CDCompletionContext context)
+      throws CompletionException {
     if (concreteCD.getCDClassesList().stream()
         .noneMatch(cClass -> context.getTypeIncStrategy().isMatched(cClass, referenceType))) {
       addTypeIncarnation(concreteCD, referenceType);
@@ -21,7 +23,8 @@ public class BaseTypeInCDCompleter extends AbstractTypeInCDCompleter {
 
   @Override
   protected void completeType(
-      ASTCDDefinition concreteCD, ASTCDInterface referenceType, CDCompletionContext context) {
+      ASTCDDefinition concreteCD, ASTCDInterface referenceType, CDCompletionContext context)
+      throws CompletionException {
     if (concreteCD.getCDInterfacesList().stream()
         .noneMatch(
             cInterface -> context.getTypeIncStrategy().isMatched(cInterface, referenceType))) {
@@ -32,7 +35,8 @@ public class BaseTypeInCDCompleter extends AbstractTypeInCDCompleter {
 
   @Override
   protected void completeType(
-      ASTCDDefinition concreteCD, ASTCDEnum referenceType, CDCompletionContext context) {
+      ASTCDDefinition concreteCD, ASTCDEnum referenceType, CDCompletionContext context)
+      throws CompletionException {
     if (concreteCD.getCDEnumsList().stream()
         .noneMatch(cEnum -> context.getTypeIncStrategy().isMatched(cEnum, referenceType))) {
       addTypeIncarnation(concreteCD, referenceType);
